requirements.txt
selenium
beautifulsoup4
python-docx
